ex00 unsigned int

+++ c02/ex00/ft_strcpy.c
@@ -6,13 +6,13 @@
 /*   By: msloot <msloot@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 12:20:10 by msloot            #+#    #+#             */
-/*   Updated: 2023/09/14 10:27:35 by msloot           ###   ########.fr       */
+/*   Updated: 2023/09/14 15:36:54 by msloot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char	*ft_strcpy(char *dest, char *src)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (src[i] != '\0')
@@ -25,7 +25,6 @@ char	*ft_strcpy(char *dest, char *src)
 }
 
 /*
-
 #include <stdio.h>
 
 int	main(void)
